backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
mysqlclient==2.2.7
requests==2.32.4
python-dotenv==1.1.1
django-autoslug==1.9.9
python-jose==3.5.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
setuptools==80.9.0
gunicorn==23.0.0
